# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pantheon Template Service API",
        "description": "A comprehensive Laravel API with React frontend, health monitoring, and system management",
        "contact": {
            "name": "Pantheon Development Team",
            "email": "admin@pantheon.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Development Server"
        },
        {
            "url": "https://api.pantheon.com",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/v1/system/status": {
            "get": {
                "tags": [
                    "System"
                ],
                "summary": "Get comprehensive system status",
                "description": "Returns detailed system status including all health checks, system information, and recommendations",
                "operationId": "getSystemStatus",
                "responses": {
                    "200": {
                        "description": "System status retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SystemStatusResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve system status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to retrieve system status"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/system/quick-status": {
            "get": {
                "tags": [
                    "System"
                ],
                "summary": "Get quick system status",
                "description": "Returns quick system status optimized for dashboard cards with essential metrics",
                "operationId": "getQuickStatus",
                "responses": {
                    "200": {
                        "description": "Quick status retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/QuickStatusResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve quick status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to retrieve quick status"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SystemStatusResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Success"
                    },
                    "data": {
                        "properties": {
                            "status": {
                                "type": "string",
                                "example": "healthy"
                            },
                            "message": {
                                "type": "string",
                                "example": "All V1 systems operational"
                            },
                            "timestamp": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "version": {
                                "type": "string",
                                "example": "1.0.0"
                            },
                            "api_version": {
                                "type": "string",
                                "example": "v1"
                            },
                            "checks": {
                                "type": "object"
                            },
                            "system_info": {
                                "type": "object"
                            },
                            "recommendations": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        },
                        "type": "object"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "QuickStatusResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Success"
                    },
                    "data": {
                        "properties": {
                            "status": {
                                "type": "string",
                                "example": "healthy"
                            },
                            "uptime": {
                                "type": "string",
                                "example": "13:37  up 23:29, 2 users, load averages: 3.75 3.44 4.08"
                            },
                            "memory_usage": {
                                "type": "number",
                                "format": "float",
                                "example": 1.56
                            },
                            "disk_usage": {
                                "type": "number",
                                "format": "float",
                                "example": 63.43
                            },
                            "database_status": {
                                "type": "string",
                                "example": "connected"
                            },
                            "cache_status": {
                                "type": "string",
                                "example": "working"
                            },
                            "timestamp": {
                                "type": "string",
                                "format": "date-time"
                            }
                        },
                        "type": "object"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}